Datatset of the design for the data attentionshifting_signal

Design of an experiment measuring the EEG brain activity of 15 participants who have been shown images of neutral and angry faces. Those faces were shown at a different visibility 16ms and 166ms and were displayed either to the left or to the right of a screen. The laterality, sex, age, and 2 measures of anxiety for each subjects are also available. The amplitude of the EEG recording are located in the dataset attentionshifting_signal.

id : identifier of the subject.

visibility : time of exposure to the image (16ms: subliminal or 166ms:supraliminal).

emotion : type of emotion of the image (angry or neutral).

direction : position of image one the screen (left or right).

laterality_id : measure of laterality of the subject.

age : age of the subject.

sex : sex of the subject.

STAIS_state : measure of the state of anxiety of the subject.

STAIS_trait : measure of the personality trait of anxiety of the subject.

data(attentionshifting_design)

Format

A data frame with 120 rows and 10 variables.

See also