Dataset of a control study in psychology.

A subset of a control experiment measuring the impulsive approach tendencies toward physical activity or sedentary behaviors.

data(jpah2016)

Format

A data frame with 38 rows and 8 variables.

Details

id identifier of the subject.

bmi body mass index.

age.

sex.

condition the experimental condition where the task was to approach physical activity and avoid sedentary behavior (ApSB_AvPA), approach sedentarity behavior and avoid physical activity (ApPA_AvSB), and a control condition (control).

time pre, post.

iapa measure of impulsive approach tendencies toward physical activity (dependant variable).

iasb measure of impulsive approach tendencies toward sedentary behavior (dependant variable).

References

Cheval, B., Sarrazin, P., Pelletier, L., & Friese, M. (2016). Effect of retraining approach-avoidance tendencies on an exercise task: A randomized controlled trial. Journal of Physical Activity and Health, 13(12), 1396-1403.

See also