Summarize of a clusterlm object.

Display the corrected p-values for each effects.

# S3 method for clusterlm
summary(
  object,
  alternative = "two.sided",
  multcomp = NULL,
  table_type = NULL,
  ...
)

Arguments

object
A clusterlm object.

alternative
A character string indicating the alternative hypothesis. Choose between "two.sided", "greater", "less". Default is "two.sided".

multcomp
A character string indicating the multiple comparison procedure to display.

table_type
A character string indicating the type of table to display. Choose between "cluster", which aggregates test into pseudo-clusters (see details for the interpretations) or "full" which displays the p-values for all tests. See details for default values.

...
Further arguments see details.

Value

A table for each effect indicating the statistics and p-values of the clusters.

Details

It creates the full table when the number of tests is <=15 and creates a table of pseudo-clusters overwise. Note that for the "troendle" method is not based on clustering of the data and the table of pseudo-clusters should only be used to facilitate the reading of the results.

See also