Plot cluster or parameters.

Plot method for class clusterlm.

# S3 method for clusterlm
plot(
  x,
  effect = "all",
  type = "statistic",
  multcomp = x$multcomp[1],
  alternative = "two.sided",
  enhanced_stat = FALSE,
  nbbaselinepts = 0,
  nbptsperunit = 1,
  distinctDVs = NULL,
  ...
)

Arguments

x

A clusterlm object.

effect

A vector of character naming the effects to display. Default is "all".

type

A character string that specified the values to highlight. "statistic" or "coef" are available. Default is "statistic".

multcomp

A character string specifying the method use to correct the p-value. It should match the one computed in the object. Default is the (first) method in the call to clusterlm. See clusterlm.

alternative

A character string specifying the alternative hypothesis for the t-test. The available options are "greater", "less" and "two.sided". Default is "two.sided".

enhanced_stat

A logical. Default is FALSE. If TRUE, the enhanced statistic will be plotted overwise it will plot the observed statistic. Change for the "tfce" or the "clustermass" multiple comparisons procedures.

nbbaselinepts

An integer. Default is 0. If the origin of the x axis should be shifted to show the start of the time lock, provide the number of baseline time points.

nbptsperunit

An integer. Default is 1. Modify this value to change the scale of the label from the number of points to the desired unit. If points are e.g. sampled at 1024Hz, set to 1024 to scale into seconds and to 1.024 to scale into milliseconds.

distinctDVs

Boolean. Should the DVs be plotted distictively, i.e. should the points be unlinked and should the name of the DVs be printed on the x axis ? Default is FALSE if the number of DV is large thant 15 or if the method is "clustermass" or "tfce".

...

further argument pass to plot.

See also